+++ htdocs/SupportedCustomerAssets.php
@@ -1,0 +1,4 @@
+<?php
+require_once("config.inc.php");
+$generator = new \CNCLTD\SupportedCustomerAssets\SupportedCustomerAssetsActiveCustomersHTMLGenerator();
+$generator->printHTML();